requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comtypes==1.4.11
contourpy==1.3.3
cryptography==46.0.4
cycler==0.12.1
distro==1.9.0
docutils==0.22.4
faiss-cpu==1.13.2
fastapi==0.128.1
fastjsonschema==2.21.2
filelock==3.20.3
fonttools==4.59.1
fsspec==2026.1.0
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
HasteContext==0.2.4
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huggingface_hub==1.3.5
id==1.5.0
idna==3.11
jaraco.classes==3.4.0
jaraco.context==6.1.0
jaraco.functools==4.4.0
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
jupyterlab_widgets==3.0.16
keyring==25.7.0
kiwisolver==1.4.9
lark==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.5
mdurl==0.1.2
mistune==3.2.0
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.6.1
nh3==0.3.2
numpy==2.3.2
openai==1.109.1
orjson==3.11.6
ormsgpack==1.12.2
packaging==25.0
pandas==2.3.1
pandocfilters==1.5.1
parso==0.8.5
pillow==11.3.0
platformdirs==4.5.1
prometheus_client==0.24.1
propcache==0.4.1
proto-plus==1.27.0
protobuf==5.29.5
psutil==7.2.2
pure_eval==0.2.3
pyarrow==23.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyjokes==0.8.3
pyparsing==3.2.3
pypiwin32==223
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-multipart==0.0.22
pyttsx3==2.99
pytz==2025.2
pywin32==311
pywin32-ctypes==0.2.3
pywinpty==3.0.2
PyYAML==6.0.3
pyzmq==27.1.0
rank-bm25==0.2.2
rdflib==7.5.0
readme_renderer==44.0
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==2.0.0
rfc3986-validator==0.1.1
rich==14.3.1
rpds-py==0.30.0
safetensors==0.7.0
-e git+https://github.com/Swastika-311/scaledown@43826532367e868157a37956a5b036346351b5d9#egg=scaledown
scikit-learn==1.8.0
scipy==1.17.0
Send2Trash==2.1.0
sentence-transformers==5.2.2
sentry-sdk==2.51.0
setuptools==80.10.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8.3
starlette==0.50.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.11.0
tinycss2==1.4.0
tokenizers==0.22.2
torch==2.10.0
tornado==6.5.4
tqdm==4.67.2
traitlets==5.14.3
transformers==5.0.0
tree-sitter==0.25.2
tree-sitter-c-sharp==0.23.1
tree-sitter-embedded-template==0.25.0
tree-sitter-language-pack==0.13.0
tree-sitter-yaml==0.7.2
twine==6.2.0
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.6.2
uuid_utils==0.14.0
uvicorn==0.40.0
wcwidth==0.5.3
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
wheel==0.46.3
widgetsnbextension==4.0.15
xxhash==3.6.0
zstandard==0.25.0
